feedreader/requirements.txt
feedparser
azure-storage-blob
azure-identity
